+++ client/src/views/BudgetDisplay.vue
@@ -1,8 +1,6 @@
 <template>
     <b-container>
-        <p class="red">{{message}}</p>
         <h1>Your Budget</h1>
-        <b-button variant="danger" v-on:click="deleteBudgetCollection">Remove Kebab</b-button>
         <b-row align-h="center">
             <b-col cols="12" sm="6" md="4" v-for="budget in budgets" v-bind:key="budget._id">
                 <budget-item class="items" v-bind:budget="budget" v-on:del-budget="deleteBudget"/>
@@ -14,6 +12,7 @@
 <script>
 import { Api } from '@/Api'
 import BudgetItem from '@/components/BudgetItem.vue'
+import cookiesC from '../cookies/cookies'
 
 export default {
   name: 'budgets',
@@ -29,9 +28,6 @@ export default {
         this.message = error.message
         this.budgets = []
       })
-      .then(() => {
-        this.message = 'testing how this works'
-      })
   },
   data() {
     return {
@@ -42,19 +38,11 @@ export default {
   },
   methods: {
     deleteBudget(id) {
-      Api.delete(`/budgets/${id}`)
+      var userId = cookiesC.getCookieValue('id')
+      Api.delete(`users/${userId}/budgets/${id}`)
         .then(response => {
           const index = this.budgets.findIndex(budget => budget._id === id)
           this.budgets.splice(index, 1)
-        })
-        .catch(error => {
-          this.message = error.message
-        })
-    },
-    deleteBudgetCollection() {
-      Api.delete('/budgets')
-        .then(response => {
-          location.reload()
         })
         .catch(error => {
           this.message = error.message
